+++ Imperium/src/Core/Lifecycle/PlayerManager.cs
@@ -52,6 +52,8 @@ internal class PlayerManager : ImpLifecycleObject
     internal bool FlyIsAscending;
     internal bool FlyIsDescending;
 
+    private static readonly int GasEmitting = Animator.StringToHash("gasEmitting");
+
     public PlayerManager(ImpBinaryBinding sceneLoaded, IBinding<int> playersConnected)
         : base(sceneLoaded, playersConnected)
     {
@@ -124,6 +126,10 @@ internal class PlayerManager : ImpLifecycleObject
     {
         player.health = 100;
         HUDManager.Instance.UpdateHealthUI(100, hurtPlayer: false);
+        HUDManager.Instance.gasHelmetAnimator.SetBool(GasEmitting, value: false);
+
+        player.bleedingHeavily = false;
+        player.criticallyInjured = false;
     }
 
     internal static void UpdateCameras(bool _) => UpdateCameras();