+++ Imperium/src/Util/Binding/ImpEvent.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Imperium.Util.Binding;
+
+public class ImpEvent
+{
+    public event Action onTrigger;
+
+    public void Trigger() => onTrigger?.Invoke();
+}